VZ 6977: Istio component cleanup (#4139)
