--- platform-operator/controllers/verrazzano/component/istio/jaeger_test.go
@@ -1,91 +1,0 @@
-// Copyright (c) 2022, Oracle and/or its affiliates.
-// Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
-
-package istio
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
-	"github.com/verrazzano/verrazzano/platform-operator/constants"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/spi"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"sigs.k8s.io/controller-runtime/pkg/client/fake"
-)
-
-var testZipkinNamespace = "foo"
-var testZipkinService = corev1.Service{
-	ObjectMeta: metav1.ObjectMeta{
-		Namespace: testZipkinNamespace,
-		Name:      "jaeger-collector",
-		Labels: map[string]string{
-			constants.KubernetesAppLabel: constants.JaegerCollectorService,
-		},
-	},
-	Spec: corev1.ServiceSpec{
-		Ports: []corev1.ServicePort{
-			{
-				Name: "foo",
-				Port: 1,
-			},
-			{
-				Name: "http-zipkin",
-				Port: 5555,
-			},
-		},
-	},
-}
-
-var expectedYaml = `
-apiVersion: install.istio.io/v1alpha1
-kind: IstioOperator
-spec:
-  values:
-    meshConfig:
-      defaultConfig:
-        tracing:
-          tlsSettings:
-            mode: "ISTIO_MUTUAL"
-          zipkin:
-            address: "jaeger-operator-jaeger-collector.verrazzano-monitoring.svc.cluster.local.:9411"
-`
-
-func TestConfigureJaeger(t *testing.T) {
-	ctxNoService := spi.NewFakeContext(fake.NewClientBuilder().WithScheme(testScheme).Build(), jaegerEnabledCR, nil, false)
-	ctxWithServiceAndUnmanagedNamespace := spi.NewFakeContext(fake.NewClientBuilder().
-		WithObjects(&testZipkinService).
-		WithScheme(testScheme).
-		Build(), jaegerEnabledCR, nil, false)
-
-	var tests = []struct {
-		name    string
-		ctx     spi.ComponentContext
-		numArgs int
-	}{
-		{
-			"2 args (tls mode and zipkin address) returned when Jaeger operator is disabled",
-			spi.NewFakeContext(fake.NewClientBuilder().Build(), &vzapi.Verrazzano{}, nil, false),
-			2,
-		},
-		{
-			"2 args (tls mode and zipkin address) returned when service is not present",
-			ctxNoService,
-			2,
-		},
-		{
-			"2 args (tls mode and zipkin address) returned when service is present",
-			ctxWithServiceAndUnmanagedNamespace,
-			2,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			yamlString, err := configureJaegerTracing()
-			assert.NoError(t, err)
-			assert.YAMLEq(t, expectedYaml, yamlString)
-		})
-	}
-}
