+++ platform-operator/manifests/profiles/v1alpha1/base.yaml
@@ -126,6 +126,14 @@ spec:
             apiVersion: install.istio.io/v1alpha1
             kind: IstioOperator
             spec:
+              values:
+                meshConfig:
+                  defaultConfig:
+                    tracing:
+                      tlsSettings:
+                        mode: ISTIO_MUTUAL
+                      zipkin:
+                        address: jaeger-operator-jaeger-collector.verrazzano-monitoring.svc.cluster.local.:9411
               components:
                 pilot:
                   k8s:
